Return raw EINTR from epoll_wait and fix TestRunOneFor
epoll_wait can be interrupted by some signal, which most of the time
is SIGURG, sent by the scheduler to preempt the current goroutine. This
happens in the modified test due to the fact the goroutine does not
do anything for too long of a time, resulting in the scheduler trying
to put it (and hence the process) to sleep.

Two things:
- we first forward the EINTR to IO and convert that to an ErrTimeout
- we reduce the polling time for epoll_wait to 5ms from 100ms in order
to not get interrupted by SIGURG. By default, the runtime will preempt
a goroutine when it is running for more than 10ms.

link: https://github.com/golang/go/issues/37942
link: https://medium.com/a-journey-with-go/go-asynchronous-preemption-b5194227371c
link: https://github.com/golang/proposal/blob/master/design/24543-non-cooperative-preemption.md