Remove StateInactive from WebsocketStream.
StateInactive was introduced in order to prevent a handshake whilst one
is already ongoing. It is now removed in order to have a cleaner reset
sequence and to correctly place the stream in StateTerminated if the
handshake fails.

Preventing a handshake whilst one is already ongoing is left to the
client.