Add Linux syscall.RawSyscall and syscall.Syscall benchmark
We initially used syscall.RawSyscall on Linux to avoid potentially
spawning a thread in the hotpath. But that is not safe if you do not use
sonic exclusively and the syscall is blocking. The solution is to use
syscall.Syscall which also tells the runtime that the syscall might
block - the goroutine will then not be able to starve off other
goroutines from running. This is important when collecting garbage - no
GC goroutine should be starved.

This commit just adds two benchmarks comparing syscall.Syscall to
syscall.RawSyscall.