SlotSequencer can now offset slots
... through prefix sums maintained by a FenwickTree.

Why do we need to offset Slots? Consider the following slots:
- (0, 1), (1, 1), (2, 1). The first digit is the Index, second is the
  Length.
- These slots live in the saved area of the ByteBuffer already
- Say we Discard (1, 1). That means we are left with (0, 1) and (2, 1).
  But the Discard already copied (2, 1) over (1, 1), so (2, 1) is
  actually (1, 1) in the buffer.
- So (2, 1) needs to be offset by (1, 1)s Length which is 1.

Given the above, this means that if we maintain our own list of Slots
and we start discarding those slots from the ByteBuffer we need to take
care such that any slots that follow a popped slot are offset by the
popped slot's length. The slots that precede the popped slot don't need
any offsetting.

A not so effcient solution would be to iterate over all slots following
a popped slot and update their Index like:
- Index = Index -poppedSlot.Length.

A more efficient solution is to only leave this offsetting to be done
just before the discard. That means we must accumulate offsets in some
effcient way, and apply those accumulated offsets to the Slot before
discarding it. That efficient way is a FenwickTree, which can give us
prefix sums over a range really fast.