Handle empty close frames correctly
Per the websocket spec, a close frame *MAY* contain some bytes
indicating the status and reason for closing the connection. The status
will occupy 2 bytes and the reason the rest of the space allowed to a
control frame.

Emphasis on *MAY*: this means we must check the close frame has a
payload before starting to decode it, otherwise we will trigger a
runtime panic by accessing a slice outside it's length.

If no status is given, the websocket spec tells us to return
CloseNoStatus=1005, without a reason.