Pending reads/writes container is now an array
This removes the random heap access, allocations and hash computations
we have by using a map. One step closer to running sonic with zero
allocations and GC off.

This is possible because all processes have a limit to the number file
descriptors. A new file descriptor returned by the kernel will be in the
range [0, limit) where limit = 1024 by default. This limit can be
increased up to a "hard-limit" by unprivileged users. The hard limit is
around 500k. I find it hard to believe anybody in trading would want to
set the limit so high, let alone make it infinite with RLIMIT_INFINITY.
This means the pendingReads/pendingWrites arrays will together occupy at
most 1MB in the worst case. 8KB by default, which covers the 1024 limit.
Good tradeoff for performance and a zero-alloc library.

Code points from the kernel:
- any new fd is installed here: https://github.com/torvalds/linux/blob/5939d45155bb405ab212ef82992a8695b35f6662/fs/file.c#L1322
- the limit is RLIMIT_NOFILE: https://github.com/torvalds/linux/blob/5939d45155bb405ab212ef82992a8695b35f6662/fs/file.c#L564
- we read the soft limit from RLIMIT_NOFILE here: https://github.com/torvalds/linux/blob/master/include/linux/sched/signal.h#L766
- this soft limit is 1024 by default