Merge pending reads with pending writes. Robustly handle the degenerate case.
Having separate containers for pending reads and writes in IO didn't make much sense. If an object goes out of scope, it doesn't matter whether it has a scheduled read or write in progress. The only thing that matters is that there's an async operation in progress and the event loop will need the object's memory when that operation completes. The pending container takes care of that. The only constraint here is that the programmer should ensure that an object taking an IO by reference should only own a single PollData object.

Per @grddev's suggestion, we now cover the degenerate case where the application has more than 4k file descriptors with a map. This ensures we don't crash the machine with OOM if somebody sets RLIMIT_NOFILE to infinity for an application that creates a large amount of file descriptors.