Issue-41:
In file.go, restored most of it back to original version. The file struct is intact and the referenced queue is now in the new construct qConn which behaves as the AsyncWriter.
Refactored body of method scheduleWrite into the more generic scheduleWriteFunc, so as to define the handler func.
In queue.go, changed from using the conn struct to the new qConn.
Created qConn, as the composition of file and queue.